Improving network analysis

+++ network-export.sh
@@ -4,10 +4,12 @@ OUTPUT_DIR=$1
 
 download() {
   NAME=$1
-  echo "downloading ${NAME}.csv ..."
-  cat ${OUTPUT_DIR}/${NAME}.csv.dir/part-* > ${OUTPUT_DIR}/${NAME}.csv
-  rm -rf ${OUTPUT_DIR}/${NAME}.csv.dir
-  wc -l ${OUTPUT_DIR}/${NAME}.csv
+  if [[ -d ${OUTPUT_DIR}/${NAME}.csv.dir ]]; then
+    echo "downloading ${NAME}.csv ..."
+    cat ${OUTPUT_DIR}/${NAME}.csv.dir/part-* > ${OUTPUT_DIR}/${NAME}.csv
+    rm -rf ${OUTPUT_DIR}/${NAME}.csv.dir
+    wc -l ${OUTPUT_DIR}/${NAME}.csv
+  fi
 }
 
 download network-nodes-indegrees
@@ -19,4 +21,5 @@ download network-nodes-components
 download network-nodes-components-stat
 download network-nodes-components-histogram
 download network-nodes-degrees
-download network-nodes-degree-stat
+download network-nodes-degrees-stat
+download network-nodes-degrees-histogram
